requirements.txt
discord.py==1.7.3
mcstatus==8.0.0